+++ R/ssd-plot.R
@@ -345,7 +345,10 @@ ssd_plot <- function(
   }
 
   gp <- gp +
-    theme(text = element_text(size = text_size))
+    theme(
+      text = element_text(size = text_size),
+      axis.text.x = ssd_element_text_hc()
+    )
 
   gp
 }